Move client common code to root module
